+++ framework/domain/src/main/java/is/codion/framework/domain/entity/EntityDefinition.java
@@ -255,7 +255,7 @@ public interface EntityDefinition {
      * @param conditionProvider the condition provider
      * @return this {@link Builder} instance
      */
-    Builder conditionProvider(ConditionType conditionType, ConditionProvider conditionProvider);
+    Builder condition(ConditionType conditionType, ConditionProvider conditionProvider);
 
     /**
      * @param backgroundColorProvider the background color provider