+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/SelectionDialogBuilder.java
@@ -61,10 +61,10 @@ public interface SelectionDialogBuilder<T> extends DialogBuilder<SelectionDialog
   Collection<T> select();
 
   /**
-   * Provides the user with the ability to select a value.
+   * Provides the user with the ability to select a single value.
    * @param <T> the value type
    */
-  interface Selector<T> {
+  interface SingleSelector<T> {
 
     /**
      * @param dialogOwner the dialog owner
@@ -72,5 +72,19 @@ public interface SelectionDialogBuilder<T> extends DialogBuilder<SelectionDialog
      * @throws is.codion.common.model.CancelException in case the user cancelled
      */
     Optional<T> select(JComponent dialogOwner);
+  }
+
+  /**
+   * Provides the user with the ability to select one or more values.
+   * @param <T> the value type
+   */
+  interface MultiSelector<T> {
+
+    /**
+     * @param dialogOwner the dialog owner
+     * @return the selected values, an empty Collections if nothing was selected
+     * @throws is.codion.common.model.CancelException in case the user cancelled
+     */
+    Collection<T> select(JComponent dialogOwner);
   }
 }