+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/dialog/DialogsTest.java
@@ -31,7 +31,6 @@ import org.junit.jupiter.api.Test;
 
 import javax.swing.JLabel;
 import javax.swing.JPanel;
-import javax.swing.JTextField;
 import java.awt.event.KeyEvent;
 import java.util.Collections;
 
@@ -174,10 +173,5 @@ public final class DialogsTest {
   void selectionDialogNonExistingDefaultSelection() {
     assertThrows(IllegalArgumentException.class, () -> Dialogs.selectionDialog(Collections.singletonList("helloist"))
             .defaultSelection("hello"));
-  }
-
-  @Test
-  void createBrowseAction() {
-    Dialogs.createBrowseAction(new JTextField());
   }
 }