+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityPanel.java
@@ -22,6 +22,7 @@ import is.codion.common.Configuration;
 import is.codion.common.event.Event;
 import is.codion.common.i18n.Messages;
 import is.codion.common.property.PropertyValue;
+import is.codion.common.state.State;
 import is.codion.common.value.Value;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.EntityType;
@@ -127,14 +128,6 @@ public class EntityPanel extends JPanel {
    */
   public static final PropertyValue<Boolean> INCLUDE_TOGGLE_EDIT_PANEL_CONTROL =
           Configuration.booleanValue("is.codion.swing.framework.ui.EntityPanel.includeToggleEditPanelControl", true);
-
-  /**
-   * Specifies whether actions to hide detail panels or show them in a dialog are available to the user<br>
-   * Value type: Boolean<br>
-   * Default value: true
-   */
-  public static final PropertyValue<Boolean> INCLUDE_DETAIL_PANEL_CONTROLS =
-          Configuration.booleanValue("is.codion.swing.framework.ui.EntityPanel.includeDetailPanelControls", true);
 
   /**
    * Specifies whether the edit controls (Save, update, delete, clear, refresh) should be on a toolbar instead of a button panel<br>
@@ -269,6 +262,7 @@ public class EntityPanel extends JPanel {
   private final DetailController detailController;
   private final Value<String> caption;
   private final Value<PanelState> editPanelState = Value.value(EMBEDDED, EMBEDDED);
+  private final State disposeEditDialogOnEscape = State.state(DISPOSE_EDIT_DIALOG_ON_ESCAPE.get());
 
   private final Settings settings = new Settings();
 
@@ -276,7 +270,6 @@ public class EntityPanel extends JPanel {
   private EntityPanel parentPanel;
   private EntityPanel previousSiblingPanel;
   private EntityPanel nextSiblingPanel;
-  private boolean disposeEditDialogOnEscape = DISPOSE_EDIT_DIALOG_ON_ESCAPE.get();
 
   private boolean initialized = false;
 
@@ -647,19 +640,11 @@ public class EntityPanel extends JPanel {
   }
 
   /**
-   * @return true if the edit dialog is disposed of on ESC
+   * @return the State controlling whtether the edit dialog is disposed of on ESC
    * @see EntityPanel#DISPOSE_EDIT_DIALOG_ON_ESCAPE
    */
-  public final boolean isDisposeEditDialogOnEscape() {
+  public final State disposeEditDialogOnEscape() {
     return disposeEditDialogOnEscape;
-  }
-
-  /**
-   * @param disposeEditDialogOnEscape if true then the edit dialog is disposed of on ESC
-   * @see EntityPanel#DISPOSE_EDIT_DIALOG_ON_ESCAPE
-   */
-  public final void setDisposeEditDialogOnEscape(boolean disposeEditDialogOnEscape) {
-    this.disposeEditDialogOnEscape = disposeEditDialogOnEscape;
   }
 
   /**
@@ -1137,7 +1122,7 @@ public class EntityPanel extends JPanel {
             .owner(this)
             .title(caption.get())
             .modal(false)
-            .disposeOnEscape(disposeEditDialogOnEscape)
+            .disposeOnEscape(disposeEditDialogOnEscape.get())
             .onClosed(windowEvent -> editPanelState.set(HIDDEN))
             .build();
   }